I/O: Harden geometry import for invalid data
* Use int64_t to avoid overflows, following our C++ guidelines.
* Add various index range checks for code that runs before mesh
  validation.
* Match some loop iteration variables to the size type.
* Ensure number of elements fit into int where needed.
* OBJ vertex weight import issue with || true (debug leftover?).
* USD invalid shape key size check (> instead of >=).
* USD now preserves vertex positions for some invalid topology,
  this the reason for usd_mesh_custom.txt test update.
* Note FBX import has fewer checks because ufbx already handles most.

Some of these are defensive, some prevent actual crashes. Tests were
added for various crashes.

These checks are meant to run in combination with mesh validation. We
try to prevent out of bounds access before validation but leave invalid
topology detection to the validation code.

Pull Request: https://projects.blender.org/blender/blender/pulls/158307

+++ source/blender/io/ply/importer/ply_import_data.cc
@@ -412,7 +412,7 @@ static const char *load_face_element(PlyReadBuffer &file,
     return "Face element vertex indices property must be a list";
   }
 
-  data->face_vertices.reserve(element.count * 3);
+  data->face_vertices.reserve(int64_t(element.count) * 3);
   data->face_sizes.reserve(element.count);
 
   if (header.type == PlyFormatType::ASCII) {
